+++ framework/db-local/src/main/java/is/codion/framework/db/local/DefaultLocalEntityConnection.java
@@ -985,7 +985,7 @@ final class DefaultLocalEntityConnection implements LocalEntityConnection {
 							.collect(toList()));
 		}
 
-		if (referencedKey.primaryKey()) {
+		if (referencedKey.primary()) {
 			return primaryKeyMap(referencedEntities);
 		}