+++ framework/domain/src/main/java/is/codion/framework/domain/entity/DefaultKeyBuilder.java
@@ -29,11 +29,11 @@ final class DefaultKeyBuilder implements Entity.Key.Builder {
 	private final EntityDefinition definition;
 	private final Map<Column<?>, Object> values = new HashMap<>();
 
-	private boolean primaryKey = true;
+	private boolean primary = true;
 
 	DefaultKeyBuilder(Entity.Key key) {
-		this(key.entityDefinition());
-		this.primaryKey = key.primaryKey();
+		this(key.definition());
+		this.primary = key.primary();
 		key.columns().forEach(column -> values.put(column, key.get(column)));
 	}
 
@@ -45,7 +45,7 @@ final class DefaultKeyBuilder implements Entity.Key.Builder {
 	public <T> Entity.Key.Builder with(Column<T> column, T value) {
 		ColumnDefinition<T> columnDefinition = definition.columns().definition(column);
 		if (!columnDefinition.primaryKey()) {
-			primaryKey = false;
+			primary = false;
 		}
 		values.put(column, value);
 
@@ -54,11 +54,11 @@ final class DefaultKeyBuilder implements Entity.Key.Builder {
 
 	@Override
 	public Entity.Key build() {
-		return new DefaultKey(definition, initializeValues(new HashMap<>(values)), primaryKey);
+		return new DefaultKey(definition, initializeValues(new HashMap<>(values)), primary);
 	}
 
 	private Map<Column<?>, Object> initializeValues(Map<Column<?>, Object> values) {
-		if (primaryKey && !values.isEmpty()) {
+		if (primary && !values.isEmpty()) {
 			//populate any missing primary key attributes with null values,
 			//DefaultKey.equals() relies on the key attributes being present
 			definition.primaryKey().columns().forEach(attribute -> values.putIfAbsent(attribute, null));